# perf: replace assert with debug_assert for data validation